Chore: refactor IAM Org layout and fix docs
