+++ docs/resources/iam_org.md
@@ -35,6 +35,9 @@ The following arguments are supported:
 * `name` - (Required) The name of the Org in IAM
 * `description` - (Required) The description of the Org
 * `parent_org_id` - (Required if not root org) The parent Org ID (GUID)
+* `display_name` - (Optional) The name of the organization suitable for display.
+* `type` - (Optional) The type of the organization e.g. `Hospital` 
+* `external_id` - (Optional)  Identifier defined by client which identifies the organization on the client side
 * `is_root_org` - (Optional) Marks the Org as a root organization (boolean)
 * `wait_for_delete` - (Optional) Blocks until the organization delete has completed. Default: `false`.
   The organization delete process can take some time as all its associated resources like
