+++ internal/services/iam/organization/resource_iam_organization.go
@@ -1,4 +1,4 @@
-package iam
+package organization
 
 import (
 	"context"
